pyproject.toml
[tool.poetry]
name = "checkbox-reports-api"
version = "0.1.0"
description = "API для отримання та обробки звітів з Checkbox.ua"
authors = ["Your Name <your.email@example.com>"]

packages = [
    { include = "src" }
]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.1.1"
sqlalchemy = "^2.0.20"
alembic = "^1.11.3"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
email-validator = "^2.0.0"
psycopg2-binary = "^2.9.7"
pydantic-settings = "^2.8.1"
APScheduler = "^3.10.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
loguru = "^0.7.3"
jinja2 = "^3.1.6"
fastapi-mail = "^1.4.2"
aiosqlite = "^0.21.0"
aiogram = "^3.20.0.post0"
supabase = "^2.15.1"
redis = "^6.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
